feat(settings, electronic woodfish); fix(many things).

+++ Firmware/ZeptoWatch/ZeptoWatch/Drivers/battery/battery.h
@@ -22,7 +22,7 @@ void Battery_StopSampling();
 
 int Battery_IsCharging();
 uint16_t Battery_GetADCData();
-uint16_t Battery_GetPowerPercentage();
+int Battery_GetPowerPercentage();
 
 #ifdef __cplusplus
 }
